requirements_lightning.txt
torch>=1.13.0
torchvision>=0.14.0
pytorch-lightning>=2.0.0
lightning>=2.0.0
numpy>=1.21.0
tqdm>=4.64.0
tensorboard>=2.9.0
wandb>=0.13.0
opencv-python>=4.6.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
h5py>=3.7.0
scipy>=1.8.0
torchmetrics>=0.11.0
hydra-core>=1.3.0
omegaconf>=2.3.0
